fix: Switch to comply with best practices
