+++ package.json
@@ -45,8 +45,8 @@
 		"README.md"
 	],
 	"dependencies": {
+		"@cheqd/did-provider-cheqd": "^3.6.4",
 		"@cheqd/sdk": "^3.6.1",
-		"@cheqd/did-provider-cheqd": "^3.6.1",
 		"@cosmjs/amino": "^0.31.0",
 		"@cosmjs/encoding": "^0.31.0",
 		"@logto/express": "^2.0.2",
