+++ src/services/connectors/verida.ts
@@ -1,12 +1,12 @@
-import { EnvironmentType } from '@verida/types'
+import type { EnvironmentType } from '@verida/types'
 import { Context, Network } from '@verida/client-ts'
 import { AutoAccount } from '@verida/account-node'
 
-import { CredentialDataRecord, DataRecord } from '../../types/verida.js'
+import type { CredentialDataRecord, DataRecord } from '../../types/verida.js'
 import { VERIDA_APP_NAME, VERIDA_CREDENTIAL_RECORD_SCHEMA } from '../../types/constants.js'
 
 import * as dotenv from 'dotenv'
-import { VerifiableCredential } from '@veramo/core'
+import type { VerifiableCredential } from '@veramo/core'
 dotenv.config()
 
 const { VERIDA_NETWORK, POLYGON_RPC_URL, VERIDA_PRIVATE_KEY, POLYGON_PRIVATE_KEY } = process.env
