+++ src/services/identity/unauthorized.ts
@@ -1,13 +1,16 @@
 import { LocalIdentity } from './local.js';
-import { VerifiableCredential, IVerifyResult, VerifiablePresentation } from '@veramo/core';
+import type {
+	VerifiableCredential,
+	IVerifyResult,
+	VerifiablePresentation,
+} from '@veramo/core';
 import { MemoryPrivateKeyStore } from '@veramo/key-manager';
 import { KeyManagementSystem } from '@veramo/kms-local';
-import { StatusCheckResult } from '@cheqd/did-provider-cheqd/build/types/agent/ICheqd';
 
-import { CheckStatusListOptions, VerificationOptions, DefaultRPCUrl } from '../../types/types.js';
+import { type CheckStatusListOptions, type VerificationOptions, DefaultRPCUrl } from '../../types/shared.js';
 import { Connection } from '../../database/connection/connection.js';
 import { Veramo } from './agent.js';
-import { CheqdDIDProvider } from '@cheqd/did-provider-cheqd';
+import { CheqdDIDProvider, type StatusCheckResult } from '@cheqd/did-provider-cheqd';
 import { CheqdNetwork } from '@cheqd/sdk';
 import * as dotenv from 'dotenv';
 
@@ -58,6 +61,7 @@ export class Unauthorized extends LocalIdentity {
 	async verifyCredential(
 		credential: VerifiableCredential | string,
 		verificationOptions: VerificationOptions,
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
 		agentId?: string
 	): Promise<IVerifyResult> {
 		return await Veramo.instance.verifyCredential(this.initAgent(), credential, verificationOptions);
@@ -66,6 +70,7 @@ export class Unauthorized extends LocalIdentity {
 	async verifyPresentation(
 		presentation: VerifiablePresentation | string,
 		verificationOptions: VerificationOptions,
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
 		agentId?: string
 	): Promise<IVerifyResult> {
 		return await Veramo.instance.verifyPresentation(this.initAgent(), presentation, verificationOptions);
@@ -74,6 +79,7 @@ export class Unauthorized extends LocalIdentity {
 	async checkStatusList2021(
 		did: string,
 		statusOptions: CheckStatusListOptions,
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
 		agentId?: string
 	): Promise<StatusCheckResult> {
 		return await Veramo.instance.checkStatusList2021(this.initAgent(), did, statusOptions);
