+++ src/types/authentication.ts
@@ -1,4 +1,4 @@
-import { Request, Response } from 'express';
+import type { Request, Response } from 'express';
 
 export enum Namespaces {
 	Testnet = 'testnet',
