+++ tsconfig.json
@@ -26,19 +26,19 @@
 
 		/* Strict Type-Checking Options */
 		"strict": true /* Enable all strict type-checking options. */,
-		// "noImplicitAny": true,                 /* Raise error on expressions and declarations with an implied 'any' type. */
+		"noImplicitAny": true,                 /* Raise error on expressions and declarations with an implied 'any' type. */
 		// "strictNullChecks": true,              /* Enable strict null checks. */
 		// "strictFunctionTypes": true,           /* Enable strict checking of function types. */
 		// "strictBindCallApply": true,           /* Enable strict 'bind', 'call', and 'apply' methods on functions. */
 		// "strictPropertyInitialization": true,  /* Enable strict checking of property initialization in classes. */
-		// "noImplicitThis": true,                /* Raise error on 'this' expressions with an implied 'any' type. */
+		"noImplicitThis": true,                /* Raise error on 'this' expressions with an implied 'any' type. */
 		// "alwaysStrict": true,                  /* Parse in strict mode and emit "use strict" for each source file. */
 
 		/* Additional Checks */
-		// "noUnusedLocals": true,                /* Report errors on unused locals. */
+		"noUnusedLocals": true,                /* Report errors on unused locals. */
 		// "noUnusedParameters": true,            /* Report errors on unused parameters. */
-		// "noImplicitReturns": true,             /* Report error when not all code paths in function return a value. */
-		// "noFallthroughCasesInSwitch": true,    /* Report errors for fallthrough cases in switch statement. */
+		"noImplicitReturns": true,             /* Report error when not all code paths in function return a value. */
+		"noFallthroughCasesInSwitch": true,    /* Report errors for fallthrough cases in switch statement. */
 		// "noUncheckedIndexedAccess": true,      /* Include 'undefined' in index signature results */
 
 		/* Module Resolution Options */
@@ -53,6 +53,8 @@
 		"esModuleInterop": true /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */,
 		// "preserveSymlinks": true,              /* Do not resolve the real path of symlinks. */
 		// "allowUmdGlobalAccess": true,          /* Allow accessing UMD globals from modules. */
+		"importsNotUsedAsValues": "error" /* Specify emit/checking behavior for imports that are only used for types */,
+		"ignoreDeprecations": "5.0" /* Ignore deprecation warnings */,
 
 		/* Source Map Options */
 		// "sourceRoot": "",                      /* Specify the location where debugger should locate TypeScript files instead of source locations. */
